qna ing2
