+++ src/main/java/com/promising/controller/ExceptionController.java
@@ -1,13 +1,11 @@
-package com.promising.controller;
-
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class ExceptionController {
-
-	@ExceptionHandler(Exception.class)
-	public String exception() {
-		return "/error";
-	}
-}
+/*
+ * package com.promising.controller;
+ * 
+ * import org.springframework.web.bind.annotation.ControllerAdvice; import
+ * org.springframework.web.bind.annotation.ExceptionHandler;
+ * 
+ * @ControllerAdvice public class ExceptionController {
+ * 
+ * @ExceptionHandler(Exception.class) public String exception() { return
+ * "/error"; } }
+ */
