+++ src/main/java/com/promising/vo/CommentVO.java
@@ -37,9 +37,6 @@ public class CommentVO {
 	private String writer;
 	@Column(nullable=false , length=4000)
 	private String contents;
-	@Column(nullable=true)
-	private Long pr_parent;
-
 	@CreationTimestamp
 	private Timestamp regDate;
 	@UpdateTimestamp
