+++ src/main/resources/templates/member/myProjectGoing.html
@@ -114,10 +114,10 @@
 											<span class="division">|</span> <span><a href="#">[[${qna.writer}]]</a></span>
 										</div>
 										<div class="pro_">
-											<a href="#">[[${pro.prTitle}]]</a>
+											<!-- <a href="#">[[${qna.prTitle}]]</a> -->
 										</div>
 										<div class="achi">
-											<span>[[${qna.regDate}]]% 달성</span>
+											<span>[[${qna.regDate}]]</span>
 										</div>
 
 									</div>
