+++ src/main/resources/templates/member/mypage.html
@@ -65,7 +65,7 @@
                                         </li>
                                          <li class="nav-item" role="presentation">
                                          
-                                        <a href="#" th:href="@{/qna/home}" class="py-2 d-block"><button class="nav-link navlinkQ p-3" id="pills-qna-tab" data-bs-toggle="pill"
+                                        <a href="#" th:href="@{/member/qna}" class="py-2 d-block"><button class="nav-link navlinkQ p-3" id="pills-qna-tab" data-bs-toggle="pill"
                                                 data-bs-target="#pills-QnA" type="button" role="tab"
                                                 aria-controls="pills-QnA" aria-selected="false">프로젝트 문의</button></a>
                                         </li>
